修复与代码优化
这次真的彻底修好xeroc套的效果了。
删除了工程内所有完全没必要的数据强转化与多余的变量。（定义变量的时候相当于申请了内存空间，删除这些东西也一定程度上能优化游戏的稳定性）
将大部分固定的变量值修改为只读，避免某些函数错误调用并修改这些变量值。

+++ Content/Items/Accessories/VoidofExtinctionLegacy.cs
@@ -84,7 +84,7 @@ namespace CalamityInheritance.Content.Items.Accessories
                         {
                             Vector2 spawn = baseSpawn;
                             spawn.X += i * 30 - (FireProjectiles * 15);
-                            Vector2 velocity = baseVelocity.RotatedBy(MathHelper.ToRadians(-FireAngleSpread / 2 + (FireAngleSpread * i / (float)FireProjectiles)));
+                            Vector2 velocity = baseVelocity.RotatedBy(MathHelper.ToRadians(-FireAngleSpread / 2 + (FireAngleSpread * i / FireProjectiles)));
                             velocity.X = velocity.X + 3 * Main.rand.NextFloat() - 1.5f;
                             int damage = (int)player.GetBestClassDamage().ApplyTo(100);
                             Projectile.NewProjectile(source, spawn, velocity, ModContent.ProjectileType<BrimstoneHellfireballFriendly2>(), damage, 5f, Main.myPlayer, 0f, 0f);
