+++ Content/Projectiles/Magic/InfernadoFriendlyold.cs
@@ -116,7 +116,7 @@ namespace CalamityInheritance.Content.Projectiles.Magic
             if (Projectile.ai[0] <= 0f)
             {
                 float swaySize = MathHelper.Pi / 30f;
-                float smolWidth = (float)Projectile.width / 5f;
+                float smolWidth = Projectile.width / 5f;
                 smolWidth *= 2f;
                 float projXChange = (float)(Math.Cos((double)(swaySize * -(double)Projectile.ai[0])) - 0.5) * smolWidth;
                 Projectile.position.X -= projXChange * -Projectile.direction;
@@ -146,7 +146,7 @@ namespace CalamityInheritance.Content.Projectiles.Magic
             Texture2D texture2D13 = Terraria.GameContent.TextureAssets.Projectile[Projectile.type].Value;
             int framing = Terraria.GameContent.TextureAssets.Projectile[Projectile.type].Value.Height / Main.projFrames[Projectile.type];
             int y6 = framing * Projectile.frame;
-            Main.spriteBatch.Draw(texture2D13, Projectile.Center - Main.screenPosition + new Vector2(0f, Projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(new Rectangle(0, y6, texture2D13.Width, framing)), Projectile.GetAlpha(lightColor), Projectile.rotation, new Vector2((float)texture2D13.Width / 2f, (float)framing / 2f), Projectile.scale, SpriteEffects.None, 0);
+            Main.spriteBatch.Draw(texture2D13, Projectile.Center - Main.screenPosition + new Vector2(0f, Projectile.gfxOffY), new Microsoft.Xna.Framework.Rectangle?(new Rectangle(0, y6, texture2D13.Width, framing)), Projectile.GetAlpha(lightColor), Projectile.rotation, new Vector2(texture2D13.Width / 2f, framing / 2f), Projectile.scale, SpriteEffects.None, 0);
             return false;
         }
     }
