+++ Content/Projectiles/Melee/Shortsword/BloodyRuptureProj.cs
@@ -49,12 +49,12 @@ namespace CalamityInheritance.Content.Projectiles.Melee.Shortsword
         {
             if (Main.rand.NextBool(5))
             {
-                int Blood = Dust.NewDust(Projectile.position + Projectile.velocity, Projectile.width, Projectile.height, DustID.Blood, (float)(Projectile.direction * 2), 0f, 15, default, 1.3f);
+                int Blood = Dust.NewDust(Projectile.position + Projectile.velocity, Projectile.width, Projectile.height, DustID.Blood, Projectile.direction * 2, 0f, 15, default, 1.3f);
             }
         }
         public override Action<Projectile> EffectBeforePullback => (proj) =>
         {
-            Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, Projectile.velocity * 6f, ModContent.ProjectileType<Blood>(), (int)(Projectile.damage * 1), Projectile.knockBack, Projectile.owner, 0f, 0f);
+            Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, Projectile.velocity * 6f, ModContent.ProjectileType<Blood>(), Projectile.damage * 1, Projectile.knockBack, Projectile.owner, 0f, 0f);
         };
 
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
