+++ Content/Projectiles/Melee/Shortsword/ExcaliburShortswordProj.cs
@@ -48,7 +48,7 @@ namespace CalamityInheritance.Content.Projectiles.Melee.Shortsword
         {
             if (Main.rand.NextBool(3))
             {
-                int Enchanted_Gold = Dust.NewDust(Projectile.position + Projectile.velocity, Projectile.width, Projectile.height, DustID.Enchanted_Gold, (float)(Projectile.direction * 2), 0f, 15, default, 1.3f);
+                int Enchanted_Gold = Dust.NewDust(Projectile.position + Projectile.velocity, Projectile.width, Projectile.height, DustID.Enchanted_Gold, Projectile.direction * 2, 0f, 15, default, 1.3f);
             }
         }
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
