+++ Content/Projectiles/Melee/Shortsword/FlameburstShortswordProj.cs
@@ -47,7 +47,7 @@ namespace CalamityInheritance.Content.Projectiles.Melee.Shortsword
         {
             if (Main.rand.NextBool(3))
             {
-                int Torch = Dust.NewDust(Projectile.position + Projectile.velocity, Projectile.width, Projectile.height, DustID.Torch, (float)(Projectile.direction * 2), 0f, 15, default, 1.3f);
+                int Torch = Dust.NewDust(Projectile.position + Projectile.velocity, Projectile.width, Projectile.height, DustID.Torch, Projectile.direction * 2, 0f, 15, default, 1.3f);
             }
         }
         public override void OnHitNPC(NPC target, NPC.HitInfo hit, int damageDone)
